+++ Controls/_searchBreadcrumbsGrid/display/SearchGridCollection.ts
@@ -6,6 +6,8 @@ import {ItemsFactory, itemsStrategy, TreeItem} from 'Controls/display';
 import BreadcrumbsItemRow from './BreadcrumbsItemRow';
 import {IOptions as ITreeGridOptions} from 'Controls/_treeGrid/display/TreeGridCollection';
 import TreeGridDataRow from 'Controls/_treeGrid/display/TreeGridDataRow';
+import CollectionEnumerator from 'Controls/_display/CollectionEnumerator';
+import {getFlatNearbyItem} from 'Controls/_display/utils/NearbyItemUtils';
 
 /**
  * Рекурсивно проверяет скрыт ли элемент сворачиванием родительских узлов
@@ -49,11 +51,6 @@ export default
    protected _$colspanBreadcrumbs: boolean;
 
    protected _$breadCrumbsMode: 'row' | 'cell';
-
-   constructor(options: IOptions<S, T>) {
-      super(options);
-      this._useFlatNearbyItemStrategy = true;
-   }
 
    protected _setupProjectionFilters(): void {
       this.addFilter(
@@ -108,6 +105,15 @@ export default
           '[Controls/_display/BreadcrumbsItem]'
       );
       this._nextVersion();
+   }
+
+   protected _getNearbyItem(
+       enumerator: CollectionEnumerator<T>,
+       item: T,
+       isNext: boolean,
+       conditionProperty?: string
+   ): T {
+      return getFlatNearbyItem(enumerator, item, isNext, conditionProperty);
    }
 
    protected _hasItemsToCreateResults(): boolean {
